Setup add_to_cart, clearing device_token, locking scheduled delivery updates, cancellation, debugger, points system

+++ lib/src/models/food_order.dart
@@ -5,7 +5,7 @@ import '../models/item.dart';
 class FoodOrder {
   String id;
   double price;
-  double quantity;
+  int quantity;
   List<Extra> extras;
   Item food;
   DateTime dateTime;
@@ -15,14 +15,14 @@ class FoodOrder {
     try {
       id = jsonMap['id'].toString();
       price = jsonMap['price'] != null ? jsonMap['price'].toDouble() : 0.0;
-      quantity = jsonMap['quantity'] != null ? jsonMap['quantity'].toDouble() : 0.0;
+      quantity = jsonMap['quantity'] != null ? jsonMap['quantity'].toInt() : 0;
       food = jsonMap['food'] != null ? Item.fromJSON(jsonMap['food']) : Item.fromJSON({});
       dateTime = DateTime.parse(jsonMap['updated_at']);
       extras = jsonMap['extras'] != null ? List.from(jsonMap['extras']).map((element) => Extra.fromJSON(element)).toList() : [];
     } catch (e) {
       id = '';
       price = 0.0;
-      quantity = 0.0;
+      quantity = 0;
       food = Item.fromJSON({});
       dateTime = DateTime(0);
       extras = [];
